add point support
